Remove ReactiveArithmeticRepository and move reactive code to Controller
- Delete ReactiveArithmeticRepository.java reactive wrapper
- Rewrite ArithmeticService as blocking/imperative service with direct JPA repository access and @Transactional
- Rewrite ArithmeticController to wrap service calls with Mono.fromCallable/Flux.fromIterable on boundedElastic
- Move recalculateMetrics() and createDefaultSettingsEntity() methods to ArithmeticService
- Fix checkstyle final variable issues in entities and mappers
